query article url by id

+++ type.go
@@ -9,5 +9,5 @@ type Article struct {
 	ID        int64  `json:"id"`
 	ChannelID int64  `json:"channelID"`
 	Url       string `json:"url"`
-	WordCount int    `json:"wordCount"`
+	WordCount int64  `json:"wordCount"`
 }
